+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## [Unreleased]
 
+## [3.6.3] (2024/07/22)
+
+- The GPT-4o mini is now available in the OpenAI Translator engine.
+- Alibaba Translate can now return the detected language.
+- Bug fixes.
+- GPT-4o mini 现已在 OpenAI 翻译引擎中可用
+- 阿里翻译现在能返回检测到的语言
+- Bug 修复
+
 ## [3.6.2] (2024/06/12)
 
 - Bug fixes.
@@ -32,7 +41,7 @@
 ## [3.5.8] (2024/04/10)
 
 - User interaction experience optimization and improvement.
-- Compatible with 2024.4.
+- Compatible with 2024.1.
 - Bug fixes.
 - 用户交互体验优化与改进
 - 兼容 IDE 2024.1 版本
@@ -487,8 +496,8 @@
 - 支持单词拆分。翻译变量名或方法名时更方便
 - Bug修复
 
-[Unreleased]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.2...HEAD
-
+[Unreleased]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.3...HEAD
+[3.6.3]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.2...v3.6.3
 [3.6.2]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.1...v3.6.2
 
 [3.6.1]: https://github.com/YiiGuxing/TranslationPlugin/compare/v3.6.0...v3.6.1